Moved Writer and friends into its own source file. Fixed a leak of the FFI-side writer if the writer did not have unwrap() called on it. Added some documentation
